+++ core/src/elements/ons-splitter/reveal-animator.js
@@ -47,11 +47,9 @@ export default class RevealSplitterAnimator extends SplitterAnimator {
     sideElement.style.left = sideElement.side === 'right' ? 'auto' : 0;
     sideElement.style.right = sideElement.side === 'right'  ? 0 : 'auto';
     sideElement.style.zIndex = 0;
-    sideElement.style.opacity = 0.9;
     sideElement.style.visibility = 'hidden';
 
     const splitter = sideElement.parentElement;
-    splitter.style.backgroundColor = 'black';
     contentReady(splitter, () => {
       if (splitter.content) {
         splitter.content.style.boxShadow = '0 0 12px 0 rgba(0, 0, 0, 0.2)';
@@ -60,11 +58,13 @@ export default class RevealSplitterAnimator extends SplitterAnimator {
   }
 
   _unsetStyles(sideElement) {
-    sideElement.style.left = sideElement.style.right = sideElement.style.zIndex = sideElement.style.opacity  = sideElement.style.visibility = '';
+    sideElement.style.left = sideElement.style.right = sideElement.style.zIndex = sideElement.style.visibility = sideElement.style.backgroundColor = '';
+    if (sideElement._content) {
+      sideElement._content.style.opacity = '';
+    }
 
     // Check if the other side needs the common styles
     if (!this._oppositeSide || this._oppositeSide.mode === 'split' || !this._oppositeSide.isOpen) {
-      sideElement.parentElement.style.backgroundColor = '';
       if (sideElement.parentElement.content) {
         sideElement.parentElement.content.style.boxShadow = '';
       }
@@ -81,14 +81,19 @@ export default class RevealSplitterAnimator extends SplitterAnimator {
     const opacity = 1 + behindDistance / 100;
 
     return {
-      transform: behindTransform,
-      opacity: opacity
+      content: {
+        opacity
+      },
+      container: {
+        transform: behindTransform
+      }
     };
   }
 
   translate(distance) {
     this.maxWidth = this.maxWidth || this._getMaxWidth();
     const menuStyle = this._generateBehindPageStyle(Math.min(distance, this.maxWidth));
+    this._side.style.backgroundColor = 'black';
     this._side.style.visibility = 'visible';
 
     if (!this._slidingElements) {
@@ -100,8 +105,10 @@ export default class RevealSplitterAnimator extends SplitterAnimator {
         .queue({
           transform: `translate3d(${this.minus + distance}px, 0px, 0px)`
         }),
+      animit(this._side._content)
+        .queue(menuStyle.content),
       animit(this._side)
-        .queue(menuStyle)
+        .queue(menuStyle.container)
     );
   }
 
@@ -130,9 +137,16 @@ export default class RevealSplitterAnimator extends SplitterAnimator {
           display: 'block'
         }),
 
+      animit(this._side._content)
+        .wait(this.delay)
+        .queue(menuStyle.content, {
+          duration: this.duration,
+          timing: this.timing
+        }),
+
       animit(this._side)
         .wait(this.delay)
-        .queue(menuStyle, {
+        .queue(menuStyle.container, {
           duration: this.duration,
           timing: this.timing
         })
@@ -168,14 +182,22 @@ export default class RevealSplitterAnimator extends SplitterAnimator {
           display: 'none'
         }),
 
+      animit(this._side._content)
+        .wait(this.delay)
+        .queue(menuStyle.content, {
+          duration: this.duration,
+          timing: this.timing
+        }),
+
       animit(this._side)
         .wait(this.delay)
-        .queue(menuStyle, {
+        .queue(menuStyle.container, {
           duration: this.duration,
           timing: this.timing
         })
         .queue(callback => {
           this._slidingElements = null;
+          this._side._content.style.opacity = this._side.style.backgroundColor = '';
           this._side.style.visibility = 'hidden';
           super.clearTransition();
           done && done();